+++ Projects/UnitTests/Source/TestIfElse.cpp
@@ -98,4 +98,32 @@ TEST_CASE("Test If/Else Branching", "[IfElse]")
 		REQUIRE(script->GetVariable("h") == true);
 	}
 
+	SECTION("Test non-standard conditional tests")
+	{
+		static const char * scriptText =
+			u8R"(
+	
+			set a to 3
+			if a
+				set a to 4
+			end
+
+			set b to []
+			if not b
+				set b to "check"
+			end
+
+			set c to null
+			if c
+				set c to 123
+			end
+
+			)";
+
+		auto script = TestExecuteScript(scriptText);
+		REQUIRE(script);
+		REQUIRE(script->GetVariable("a") == 4);
+		REQUIRE(script->GetVariable("b") == "check");
+		REQUIRE(script->GetVariable("c") == nullptr);
+	}
 }