+++ Source/Jinx.h
@@ -80,13 +80,13 @@ namespace Jinx
 	static const uint32_t MajorVersion = 0;
 
 	/// Minor version number
-	static const uint32_t MinorVersion = 11;
+	static const uint32_t MinorVersion = 12;
 
 	/// Patch number
 	static const uint32_t PatchNumber = 0;
 
 	/// Version string
-	static const char * VersionString = "0.11.0";
+	static const char * VersionString = "0.12.0";
 
 	// Forward declaration
 	class IScript;